Added the ability to calculate by blocks or resources

+++ Assets/Assets/Scripts/Calculator/CalculationResult.cs
@@ -2,7 +2,7 @@
 public struct CalculationResult
 {
     public float powerUse;
-    public int blocksUseCount;
+    public float blocksUseCount;
     public BlockDataObj recipeBlockData;
     public ResourceStack[] inputResources;
     public ResourceStack outputResource;
